+++ integration-tests/tests.py
@@ -1,6 +1,12 @@
 #!/usr/bin/python3
 
 import subprocess, argparse, shlex, sys, re, os
+from time import sleep
+
+try:
+    bob_nodes_amount_string = os.environ['BOB_NODES_AMOUNT']
+except KeyError:
+    sys.exit('Nodes amount is not set.')
 
 run_options = ['put','get','exist']
 
@@ -8,7 +14,7 @@ test_run_config = dict()
 iter = 0
 try:
     for item in run_options:
-        test_run_config[item]=str(20000+(iter % int(os.environ['BOB_NODES_AMOUNT']))) #used in get_run_args()
+        test_run_config[item]=str(20000+(iter % int(bob_nodes_amount_string))) #used in get_run_args()
         iter += 1
 except KeyError:
     sys.exit('Nodes amount is not set.')
@@ -23,6 +29,8 @@ def run_tests(behaviour, args):
         if behaviour in {'put', 'get'}:
             if f'{behaviour} errors:' in str(p) or f'panicked' in str(p):
                 sys.exit(f'{behaviour} test failed, see output')
+            if behaviour == 'put':
+                sleep(30)
         elif behaviour == 'exist':
             found_exist = re.search(r'\b[0-9]{1,}\sof\s[0-9]{1,}\b', str(p))
             if not found_exist: