+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Bob versions changelog
 - Compare vdiskid first (#594)
 - Optimize finding actual holders (#595)
 - Logger output directed to stdout instead of stderr in bobt (#651)
+- Added integration tests for aliens (#642)
+- Added integration tests for bobt (#648)
 
 #### Fixed
 - Print full error text received from Pearl in exist function (#581)